MyBatis-Plus
Former-commit-id: 2fb4ba8053eee73735e4e9580e4912bab97be3a6 [formerly 88c24f4c980ea267223079e32917d0beb7ce5763]
Former-commit-id: fd933d726c466331eddcc651fb5d94b432959491

+++ faber-admin/src/main/java/com/faber/rbac/entity/RbacRoleMenu.java
@@ -25,12 +25,12 @@ public class RbacRoleMenu extends BaseDelEntity {
 
     @ExcelProperty("ID")
     @TableId(type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     @ExcelProperty("角色ID")
-    private Integer roleId;
+    private Long roleId;
 
     @ExcelProperty("权限ID")
-    private Integer menuId;
+    private Long menuId;
 
 }
